+++ client/package.json
@@ -12,7 +12,7 @@
     "@ant-design/cssinjs": "^1.18.2",
     "@ant-design/nextjs-registry": "^1.0.2",
     "@ant-design/v5-patch-for-react-19": "^1.0.3",
-    "antd": "^5.24.7",
+    "antd": "^5.24.8",
     "axios": "^1.8.4",
     "class-transformer": "^0.5.1",
     "i18next": "^25.0.1",
