+++ client/src/app/(webRouter)/cart/page.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import Cart from "@/components/screens/card/cartScreen";
+import Cart from "@/components/screens/cart/cartScreen";
 
 const CartPage = () => {
