+++ client/src/app/(webRouter)/contact/page.tsx
@@ -1,8 +1,8 @@
+import ContactScreen from "@/components/screens/contact/view/contactScreen";
 const ContactPage = () => {
     return (
         <div>
-        <h1>Contact Us</h1>
-        <p>If you have any questions, feel free to reach out!</p>
+            <ContactScreen />
         </div>
     );
     }
